+++ Sources/VCore/Data & Domain Helpers/Image Store & Repository/Cache (Memory)/MockImageMemoryCache.swift
@@ -5,8 +5,6 @@
 //  Created by Vakhtang Kontridze on 16/5/26.
 //
 
-#if DEBUG
-
 #if canImport(UIKit)
 public import UIKit
 #elseif canImport(AppKit)
@@ -77,5 +75,3 @@ nonisolated open class MockImageMemoryCache: ImageMemoryCache, @unchecked Sendab
         type: ImageMemoryCacheCacheType
     ) {}
 }
-
-#endif