+++ Sources/VCore/Data & Domain Helpers/Image Store & Repository/Progress Cache (Memory)/MockImageProgressMemoryCache.swift
@@ -5,8 +5,6 @@
 //  Created by Vakhtang Kontridze on 16/5/26.
 //
 
-#if DEBUG
-
 #if canImport(UIKit)
 public import UIKit
 #elseif canImport(AppKit)
@@ -68,5 +66,3 @@ nonisolated open class MockImageProgressMemoryCache: ImageProgressMemoryCache, @
         cancel: Bool
     ) {}
 }
-
-#endif